+++ outlook_web/segments/01_bootstrap.py
@@ -734,7 +734,8 @@ OAUTH_SCOPES = [
     "offline_access",
     "https://graph.microsoft.com/Mail.Read",
     "https://graph.microsoft.com/Mail.ReadWrite",
-    "https://graph.microsoft.com/User.Read"
+    "https://graph.microsoft.com/User.Read",
+    "https://outlook.office.com/IMAP.AccessAsUser.All"
 ]